+++ src/pages/RepoPage/SettingsTab/tabs/GeneralTab/DangerZone/RepoState/hooks/useRepoActivation.js
@@ -1,12 +1,25 @@
 import { useUpdateRepo } from 'services/repo'
+import { useAddNotification } from 'services/toastNotification'
 
 export function useRepoActivation() {
+  const addToast = useAddNotification()
   const { mutate, ...rest } = useUpdateRepo()
 
   async function toggleRepoState(activated) {
-    mutate({
-      activated: !activated,
-    })
+    const status = activated ? 'deactivate' : 'activate'
+
+    mutate(
+      {
+        activated: !activated,
+      },
+      {
+        onError: () =>
+          addToast({
+            type: 'error',
+            text: `We were not able to ${status} this repo`,
+          }),
+      }
+    )
   }
 
   return { toggleRepoState, ...rest }
